pyproject.toml
[project]
name = "hamachirecruiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[bedrock]>=0.49.0",
    "boto3>=1.37.11",
    "daft>=0.4.7",
    "dotenv>=0.9.9",
    "fireworks-ai>=0.15.12",
    "gitpython>=3.1.44",
    "instructor[bedrock]>=1.7.4",
    "instructor>=1.7.4",
    "ipython>=9.0.2",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "pygithub>=2.6.1",
    "ray>=2.43.0",
    "pre-commit>=4.1.0",
]
